# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FileMaker\u2013Shopware Integration API",
    "version": "1.0.0",
    "description": "Production-oriented integration service reconstruction for technical and portfolio review. The Shopware source and FileMaker target are simulated/reconstructed while MongoDB Atlas is the runtime persistence layer. There is no live Shopware connection, no live FileMaker connection, and no real customer data. API v1 operational routes require x-api-key, request IDs provide correlation, and synchronization executes synchronously. Render deployment remains pending."
  },
  "servers": [
    { "url": "/", "description": "Current host" }
  ],
...
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "operationId": "listProducts",
        "summary": "List persisted products",
        "description": "Returns deterministic pagination. productNumber is exact stored-value equality; name is an escaped case-insensitive contains search.",
        "tags": ["Products"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/ProductNumberQuery" },
          { "$ref": "#/components/parameters/ProductNameQuery" },
          { "$ref": "#/components/parameters/IsActive" },
          { "$ref": "#/components/parameters/MinStock" },
          { "$ref": "#/components/parameters/ProductSort" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/ProductListSuccess" },
          "400": { "$ref": "#/components/responses/InvalidQuery" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/products/{productNumber}": {
      "get": {
        "operationId": "getProduct",
        "summary": "Read one product",
        "tags": ["Products"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/ProductNumberPath" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/ProductDetailSuccess" },
          "400": { "$ref": "#/components/responses/InvalidIdentifier" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/ProductNotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/sync/products": {
      "post": {
        "operationId": "synchronizeProducts",
        "summary": "Synchronize products",
        "description": "Synchronously imports the simulated Shopware products, creates one products-scope SyncRun, and writes one correlated SyncLog.",
        "tags": ["Synchronization"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/SynchronizationSuccess" },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "415": { "$ref": "#/components/responses/UnsupportedMediaType" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/orders": {
      "get": {
        "operationId": "listOrders",
        "summary": "List persisted orders",
        "description": "Returns deterministic pagination. orderNumber is exact stored-value equality and status is normalized to lowercase before exact matching.",
        "tags": ["Orders"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/OrderNumberQuery" },
          { "$ref": "#/components/parameters/OrderStatus" },
          { "$ref": "#/components/parameters/From" },
          { "$ref": "#/components/parameters/To" },
          { "$ref": "#/components/parameters/OrderSort" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/OrderListSuccess" },
          "400": { "$ref": "#/components/responses/InvalidQuery" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/orders/{orderNumber}": {
      "get": {
        "operationId": "getOrder",
        "summary": "Read one order",
        "tags": ["Orders"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/OrderNumberPath" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/OrderDetailSuccess" },
          "400": { "$ref": "#/components/responses/InvalidIdentifier" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/OrderNotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/sync/orders": {
      "post": {
        "operationId": "synchronizeOrders",
        "summary": "Synchronize orders",
        "description": "Synchronously imports the simulated Shopware orders, creates one orders-scope SyncRun, and writes one correlated SyncLog.",
        "tags": ["Synchronization"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/SynchronizationSuccess" },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "415": { "$ref": "#/components/responses/UnsupportedMediaType" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/sync/all": {
      "post": {
        "operationId": "synchronizeAll",
        "summary": "Synchronize products and orders",
        "description": "Synchronously runs product and order imports under exactly one all-scope SyncRun. Product, order, and dashboard logs share one runId and requestId.",
        "tags": ["Synchronization"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/SynchronizationSuccess" },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "415": { "$ref": "#/components/responses/UnsupportedMediaType" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/sync/status": {
      "get": {
        "operationId": "getSynchronizationStatus",
        "summary": "Read latest synchronization status",
        "description": "Returns the newest SyncRun for products, orders, and all scopes, or never-run when a scope has no history.",
        "tags": ["Synchronization"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/SyncStatusSuccess" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/sync/logs": {
      "get": {
        "operationId": "listSyncLogs",
        "summary": "List synchronization logs",
        "description": "Returns general history, including readable legacy records that may lack run correlation metadata.",
        "tags": ["Synchronization"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/LogEntity" },
          { "$ref": "#/components/parameters/LogStatus" },
          { "$ref": "#/components/parameters/RunIdQuery" },
          { "$ref": "#/components/parameters/From" },
          { "$ref": "#/components/parameters/To" },
          { "$ref": "#/components/parameters/LogSort" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/SyncLogListSuccess" },
          "400": { "$ref": "#/components/responses/InvalidQuery" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/sync/logs/{logId}": {
      "get": {
        "operationId": "getSyncLog",
        "summary": "Read one synchronization log",
        "tags": ["Synchronization"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/LogIdPath" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/SyncLogDetailSuccess" },
          "400": { "$ref": "#/components/responses/InvalidIdentifier" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/SyncLogNotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/sync/runs": {
      "get": {
        "operationId": "listSyncRuns",
        "summary": "List top-level synchronization runs",
        "tags": ["Synchronization"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/RunScope" },
          { "$ref": "#/components/parameters/RunStatus" },
          { "$ref": "#/components/parameters/From" },
          { "$ref": "#/components/parameters/To" },
          { "$ref": "#/components/parameters/RunSort" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/SyncRunListSuccess" },
          "400": { "$ref": "#/components/responses/InvalidQuery" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/sync/runs/{runId}": {
      "get": {
        "operationId": "getSyncRun",
        "summary": "Read one run and its logs",
        "description": "Returns logs belonging to the run in createdAt ascending order with _id ascending as the deterministic tie-breaker. startedAt remains execution lifecycle timing.",
        "tags": ["Synchronization"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/RunIdPath" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/SyncRunDetailSuccess" },
          "400": { "$ref": "#/components/responses/InvalidIdentifier" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/SyncRunNotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": { "type": "apiKey", "in": "header", "name": "x-api-key", "description": "API key required by every /api/v1 operation. Use Swagger Authorize; no key is embedded in this document." }
    },
    "headers": {
      "X-Request-Id": { "description": "Correlation ID accepted from x-request-id when safe, otherwise a generated UUID.", "schema": { "$ref": "#/components/schemas/RequestId" } }
    },
    "parameters": {
      "RequestId": { "name": "x-request-id", "in": "header", "required": false, "description": "Optional 1-100 character correlation ID using letters, digits, hyphen, underscore, period, or colon. Invalid values are replaced by a generated UUID.", "schema": { "$ref": "#/components/schemas/RequestId" } },
      "Page": { "name": "page", "in": "query", "required": false, "description": "One-based result page.", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
      "Limit": { "name": "limit", "in": "query", "required": false, "description": "Results per page.", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 25 } },
      "ProductNumberQuery": { "name": "productNumber", "in": "query", "required": false, "description": "Exact, case-sensitive stored-value equality filter; partial values do not match longer identifiers.", "schema": { "type": "string", "minLength": 1, "maxLength": 100 }, "example": "BOOK-001" },
      "ProductNameQuery": { "name": "name", "in": "query", "required": false, "description": "Escaped, regex-injection-safe, case-insensitive partial/contains search over productName.", "schema": { "type": "string", "minLength": 1, "maxLength": 100 }, "example": "API Design" },
      "IsActive": { "name": "isActive", "in": "query", "required": false, "description": "Exact active-state filter.", "schema": { "type": "boolean" } },
      "MinStock": { "name": "minStock", "in": "query", "required": false, "description": "Minimum stock quantity, inclusive.", "schema": { "type": "integer", "minimum": 0 } },
      "ProductSort": { "name": "sort", "in": "query", "required": false, "description": "Product sort; a leading minus selects descending order.", "schema": { "type": "string", "enum": ["productName", "-productName", "stockQuantity", "-stockQuantity", "syncedAt", "-syncedAt"], "default": "-syncedAt" } },
      "OrderNumberQuery": { "name": "orderNumber", "in": "query", "required": false, "description": "Exact, case-sensitive stored-value equality filter; partial values do not match longer identifiers.", "schema": { "type": "string", "minLength": 1, "maxLength": 100 }, "example": "ORD-2026-001" },
      "OrderStatus": { "name": "status", "in": "query", "required": false, "description": "Trimmed, lowercase-normalized exact order-status match.", "schema": { "type": "string", "minLength": 1, "maxLength": 50 }, "example": "open" },
      "From": { "name": "from", "in": "query", "required": false, "description": "Inclusive ISO date or datetime lower bound. Targets orderDate for orders, createdAt for SyncLogs, and startedAt for SyncRuns. Must not be later than to.", "schema": { "oneOf": [{ "type": "string", "format": "date" }, { "type": "string", "format": "date-time" }] }, "example": "2026-07-01" },
      "To": { "name": "to", "in": "query", "required": false, "description": "Inclusive ISO date or datetime upper bound. A date-only value includes the full UTC day. Must not be earlier than from.", "schema": { "oneOf": [{ "type": "string", "format": "date" }, { "type": "string", "format": "date-time" }] }, "example": "2026-07-12" },
      "OrderSort": { "name": "sort", "in": "query", "required": false, "description": "Order sort; a leading minus selects descending order.", "schema": { "type": "string", "enum": ["orderDate", "-orderDate", "amountTotal", "-amountTotal", "syncedAt", "-syncedAt"], "default": "-orderDate" } },
      "LogEntity": { "name": "entity", "in": "query", "required": false, "description": "Synchronization-log entity.", "schema": { "type": "string", "enum": ["products", "orders", "dashboard"] } },
      "LogStatus": { "name": "status", "in": "query", "required": false, "description": "Synchronization-log outcome.", "schema": { "type": "string", "enum": ["success", "failure"] } },
      "RunIdQuery": { "name": "runId", "in": "query", "required": false, "description": "Exact UUID correlation filter.", "schema": { "type": "string", "format": "uuid" } },
      "LogSort": { "name": "sort", "in": "query", "required": false, "description": "SyncLog list sort; a leading minus selects descending order.", "schema": { "type": "string", "enum": ["createdAt", "-createdAt", "durationMs", "-durationMs"], "default": "-createdAt" } },
      "RunScope": { "name": "scope", "in": "query", "required": false, "description": "Top-level synchronization scope.", "schema": { "type": "string", "enum": ["products", "orders", "all"] } },
      "RunStatus": { "name": "status", "in": "query", "required": false, "description": "SyncRun lifecycle state.", "schema": { "type": "string", "enum": ["running", "success", "failure"] } },
      "RunSort": { "name": "sort", "in": "query", "required": false, "description": "SyncRun list sort. startedAt remains the lifecycle sort field; a leading minus selects descending order.", "schema": { "type": "string", "enum": ["startedAt", "-startedAt", "durationMs", "-durationMs"], "default": "-startedAt" } },
      "ProductNumberPath": { "name": "productNumber", "in": "path", "required": true, "description": "Exact persisted product number.", "schema": { "type": "string", "minLength": 1, "maxLength": 100 }, "example": "BOOK-001" },
      "OrderNumberPath": { "name": "orderNumber", "in": "path", "required": true, "description": "Exact persisted order number.", "schema": { "type": "string", "minLength": 1, "maxLength": 100 }, "example": "ORD-2026-001" },
      "LogIdPath": { "name": "logId", "in": "path", "required": true, "description": "MongoDB ObjectId for one SyncLog.", "schema": { "type": "string", "pattern": "^[0-9a-fA-F]{24}$" }, "example": "507f1f77bcf86cd799439011" },
      "RunIdPath": { "name": "runId", "in": "path", "required": true, "description": "Public SyncRun UUID.", "schema": { "type": "string", "format": "uuid" }, "example": "123e4567-e89b-42d3-a456-426614174000" }
    },
    "responses": {
...
      "ProductListSuccess": { "description": "Product page", "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductListResponse" } } } },
      "ProductDetailSuccess": { "description": "Product detail", "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductDetailResponse" } } } },
      "OrderListSuccess": { "description": "Order page", "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrderListResponse" } } } },
      "OrderDetailSuccess": { "description": "Order detail", "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrderDetailResponse" } } } },
      "SynchronizationSuccess": { "description": "Synchronization completed synchronously", "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SynchronizationResponse" } } } },
      "SyncStatusSuccess": { "description": "Latest scope status", "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SyncStatusResponse" } } } },
      "SyncLogListSuccess": { "description": "Synchronization-log page", "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SyncLogListResponse" } } } },
      "SyncLogDetailSuccess": { "description": "Synchronization-log detail", "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SyncLogDetailResponse" } } } },
      "SyncRunListSuccess": { "description": "Synchronization-run page", "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SyncRunListResponse" } } } },
      "SyncRunDetailSuccess": { "description": "Run detail with createdAt-ordered logs", "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SyncRunDetailResponse" } } } },
      "InvalidQuery": { "description": "INVALID_QUERY", "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" }, "example": { "success": false, "error": { "code": "INVALID_QUERY", "message": "limit must be an integer from 1 to 100", "requestId": "request-42" } } } } },
      "InvalidIdentifier": { "description": "INVALID_IDENTIFIER", "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" }, "example": { "success": false, "error": { "code": "INVALID_IDENTIFIER", "message": "runId must be a valid UUID", "requestId": "request-42" } } } } },
      "InvalidRequest": { "description": "INVALID_QUERY or INVALID_REQUEST", "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "Unauthorized": { "description": "UNAUTHORIZED", "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" }, "example": { "success": false, "error": { "code": "UNAUTHORIZED", "message": "Invalid or missing API key", "requestId": "request-42" } } } } },
      "ProductNotFound": { "description": "PRODUCT_NOT_FOUND", "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "OrderNotFound": { "description": "ORDER_NOT_FOUND", "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "SyncLogNotFound": { "description": "SYNC_LOG_NOT_FOUND", "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "SyncRunNotFound": { "description": "SYNC_RUN_NOT_FOUND", "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "PayloadTooLarge": { "description": "PAYLOAD_TOO_LARGE", "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "UnsupportedMediaType": { "description": "UNSUPPORTED_MEDIA_TYPE", "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
...
      "InternalError": { "description": "INTERNAL_ERROR", "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
    },
    "schemas": {
      "RequestId": { "type": "string", "minLength": 1, "maxLength": 100, "pattern": "^[A-Za-z0-9_.:-]+$", "example": "fm-dashboard:request-42" },
...
      "ErrorDetail": {
        "type": "object", "required": ["code", "message", "requestId"],
        "properties": {
          "code": { "type": "string", "enum": ["INVALID_QUERY", "INVALID_IDENTIFIER", "INVALID_REQUEST", "PAYLOAD_TOO_LARGE", "UNSUPPORTED_MEDIA_TYPE", "UNAUTHORIZED", "PRODUCT_NOT_FOUND", "ORDER_NOT_FOUND", "SYNC_LOG_NOT_FOUND", "SYNC_RUN_NOT_FOUND", "ROUTE_NOT_FOUND", "SERVICE_NOT_READY", "INTERNAL_ERROR"] },
          "message": { "type": "string" },
          "requestId": { "$ref": "#/components/schemas/RequestId" }
        }
      },
      "ErrorResponse": { "type": "object", "required": ["success", "error"], "properties": { "success": { "type": "boolean", "enum": [false] }, "error": { "$ref": "#/components/schemas/ErrorDetail" } } },
      "Pagination": { "type": "object", "required": ["page", "limit", "total", "totalPages"], "properties": {
        "page": { "type": "integer", "minimum": 1 }, "limit": { "type": "integer", "minimum": 1, "maximum": 100 }, "total": { "type": "integer", "minimum": 0 }, "totalPages": { "type": "integer", "minimum": 0 }
      } },
      "Product": {
        "type": "object", "required": ["_id", "shopwareId", "productNumber", "productName", "stockQuantity", "isActive", "netPrice", "grossPrice", "syncedAt", "createdAt", "updatedAt"],
        "properties": {
          "_id": { "$ref": "#/components/schemas/MongoId" }, "shopwareId": { "type": "string", "example": "sw-prod-1001" }, "productNumber": { "type": "string", "example": "BOOK-001" }, "productName": { "type": "string", "example": "API Design Fundamentals" }, "stockQuantity": { "type": "number", "minimum": 0, "example": 42 }, "isActive": { "type": "boolean", "example": true }, "netPrice": { "type": "number", "minimum": 0, "example": 24.99 }, "grossPrice": { "type": "number", "minimum": 0, "example": 29.74 }, "syncedAt": { "type": "string", "format": "date-time" }, "createdAt": { "type": "string", "format": "date-time" }, "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "ProductListResponse": { "type": "object", "required": ["success", "requestId", "data", "pagination"], "properties": { "success": { "type": "boolean", "enum": [true] }, "requestId": { "$ref": "#/components/schemas/RequestId" }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } }, "pagination": { "$ref": "#/components/schemas/Pagination" } } },
      "ProductDetailResponse": { "type": "object", "required": ["success", "requestId", "data"], "properties": { "success": { "type": "boolean", "enum": [true] }, "requestId": { "$ref": "#/components/schemas/RequestId" }, "data": { "$ref": "#/components/schemas/Product" } } },
...
      "OrderListResponse": { "type": "object", "required": ["success", "requestId", "data", "pagination"], "properties": { "success": { "type": "boolean", "enum": [true] }, "requestId": { "$ref": "#/components/schemas/RequestId" }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/Order" } }, "pagination": { "$ref": "#/components/schemas/Pagination" } } },
      "OrderDetailResponse": { "type": "object", "required": ["success", "requestId", "data"], "properties": { "success": { "type": "boolean", "enum": [true] }, "requestId": { "$ref": "#/components/schemas/RequestId" }, "data": { "$ref": "#/components/schemas/Order" } } },
...
      "SyncStatusResponse": { "type": "object", "required": ["success", "requestId", "data"], "properties": {
        "success": { "type": "boolean", "enum": [true] }, "requestId": { "$ref": "#/components/schemas/RequestId" }, "data": { "type": "object", "required": ["products", "orders", "all"], "properties": { "products": { "$ref": "#/components/schemas/SyncStatusEntry" }, "orders": { "$ref": "#/components/schemas/SyncStatusEntry" }, "all": { "$ref": "#/components/schemas/SyncStatusEntry" } } }
      } },
      "SyncLogListResponse": { "type": "object", "required": ["success", "requestId", "data", "pagination"], "properties": { "success": { "type": "boolean", "enum": [true] }, "requestId": { "$ref": "#/components/schemas/RequestId" }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/SyncLog" } }, "pagination": { "$ref": "#/components/schemas/Pagination" } } },
      "SyncLogDetailResponse": { "type": "object", "required": ["success", "requestId", "data"], "properties": { "success": { "type": "boolean", "enum": [true] }, "requestId": { "$ref": "#/components/schemas/RequestId" }, "data": { "$ref": "#/components/schemas/SyncLog" } } },
      "SyncRunListResponse": { "type": "object", "required": ["success", "requestId", "data", "pagination"], "properties": { "success": { "type": "boolean", "enum": [true] }, "requestId": { "$ref": "#/components/schemas/RequestId" }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/SyncRun" } }, "pagination": { "$ref": "#/components/schemas/Pagination" } } },
      "SyncRunDetailResponse": { "type": "object", "required": ["success", "requestId", "data"], "properties": { "success": { "type": "boolean", "enum": [true] }, "requestId": { "$ref": "#/components/schemas/RequestId" }, "data": { "type": "object", "required": ["run", "logs"], "properties": { "run": { "$ref": "#/components/schemas/SyncRun" }, "logs": { "type": "array", "description": "Ascending persisted timeline by createdAt, then _id.", "items": { "$ref": "#/components/schemas/SyncLog" } } } } } },
...
      "SingleSynchronizationData": { "allOf": [{ "$ref": "#/components/schemas/SyncExecution" }, { "type": "object", "required": ["result"], "properties": { "result": { "$ref": "#/components/schemas/EntitySyncResult" } } }] },
...
      "SynchronizationResponse": { "type": "object", "required": ["success", "requestId", "data"], "properties": { "success": { "type": "boolean", "enum": [true] }, "requestId": { "$ref": "#/components/schemas/RequestId" }, "data": { "oneOf": [{ "$ref": "#/components/schemas/SingleSynchronizationData" }, { "$ref": "#/components/schemas/FullSynchronizationData" }] } } }
    }
  }
}